pyproject.toml
[tool.poetry]
name = "timestamp"
version = "0.1.0"
description = ""
authors = ["Aline Rose <aline.rose@mail.uft.edu.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
